:building_construction: refactor project structure
