+++ internal/bezel/assets/iphone-13-pro.png
