+++ pkg/frame/frame_gif.go
