+++ pkg/frame/frame_gif_test.go
@@ -3,7 +3,7 @@ package frame_test
 import (
 	"bytes"
 	approvals "github.com/approvals/go-approval-tests"
-	"github.com/nikhilhenry/xframe/frame"
+	"github.com/nikhilhenry/xframe/pkg/frame"
 	"image/gif"
 	_ "image/gif"
 	"os"
