+++ pkg/frame/frame_test.go
@@ -3,7 +3,7 @@ package frame_test
 import (
 	"bytes"
 	approvals "github.com/approvals/go-approval-tests"
-	"github.com/nikhilhenry/xframe/frame"
+	"github.com/nikhilhenry/xframe/pkg/frame"
 	"image"
 	"os"
 	"testing"
