+++ pkg/frame/testdata/frame_test.TestGenerate.it_generates_an_image_with_the_device_bezel.approved.png
